fix a test and move one more file
