+++ src/asgard_io_tests.cpp
@@ -1,6 +1,3 @@
-#include "io.hpp"
-
-#include "asgard_vector.hpp"
 #include "tests_general.hpp"
 
 using namespace asgard;
